Refactoring: move ListBuckets operations into a separate partial class of StorageClient.
If each set of logical operations has its own partial class, it'll be much easier to find the right bit of code.
An alternative would be to use extension methods for everything, but I think this approach is more straightforward.
